Add mechanism to read app info using steamcmd + registry
Using steamcmd, we can print an application's info. For the purposes of
this application, all we care about is the ufs/savefiles and
ufs/rootoverrides - this is where Steam stores the paths to the save
files.

The goal will be to create a mechanism to detect all applications
installed, iterate through them to find their save paths, reverse
engineer the config to figure out what those paths are, and generate a
games.yml file that has all this information.

The end result will be a games.yml file that will contain all app IDs
(so we can only sync that app when a game closes) and more importantly,
one that can automatically get generated with all the right stuff.
This will make this app 0-config once it's up and running.