Add mechanism to infer save file location
This is a work-in-progress: start supporting the inference of game file
locations. It introduces the model of "SteamGame" which is an
aggregation of its metadata and its config (which is read via
PrintAppInfo).

Soon, once most games are supported, I will be able to create the
games.yml based on the list of locally installed games.