Add mechanism to fetch game list from Steam
The locally installed game list is pretty isn't representative of the
games that you have on your account because it's OS-specific. It's
better if the functionality that relies on listing your games behaves
more holistically. To solve for that, we try to fetch the current user's
game list via
`https://steamcommunity.com/profiles/{steam64id}/games?xml=1`.

If a) the list can't be fetched or b) the profile is private, we'll
fall back to loading the list of installed games.