Add support for multiple save folders
The actual save format supports multiple different folders. I've been
lazy and just assuming there's just one, but this paves the way to
actually supporting everything properly. With this mechanism, we'll
start to be able to support both file patterns and multiple folders.
I've temporarily added a new interface to work with and I'll eventually
deprecate and delete the old one (getInstallationPath returning a
String).